+++ src/basic/error.ts
@@ -69,8 +69,7 @@
  * @module
  */
 
-import { isError } from "../index.js";
-import { isNumber, isString } from "./json.js";
+import { isError, isNumber, isString } from "../index.js";
 
 
 /**